Examples needed updating with new paths

+++ Examples/KelvinHelmholtz/SingleFluidIdealRandomMPIHDF5/main.cc
@@ -47,7 +47,7 @@ int main(int argc, char *argv[]) {
   ParallelEnv env(&argc, &argv, nxRanks, nyRanks, nzRanks);
 
   const int nOptionalSimArgs = 1;
-  std::vector<double> optionalSimArgs = {seed};
+  std::vector<double> optionalSimArgs = {(double)seed};
   std::vector<std::string> optionalSimArgNames = {"seed"}; 
 
   // Create an arg object that will contain all parameters needed by the simulation, that will be stored on the Data object.
